examples/healthcare/requirements.txt
fastmcp>=0.2.0
ironbook-sdk>=0.3.0
python-dotenv>=1.0.0
fastmcp-ironbook>=0.1.0

